BLACK JACK CASINO GAME
This Game consists of a player and machine plays as an opponent. There is deck of card(where King/Queen/Jack counts as 10 and ACE can be either 11 or 1. In my case, if the count is lesser than or equal to 10, i programmed in a way to consider the "Ace " as 11 and if the count is greater than 10, "Ace" will be 1.). Player has to bet and the balance is fixed to 100 in the beginning. Bank account for the player will be maintained.Player will be given 2 cards initially and machine has one card with the other card hidden. Player starts the game and continuously hits the cards until he reaches the count<=21. Once the player decided to stay and not to go on further, the machine gets the chance and starts to play. (As machine part is auto-generated, it has been fixed to stop the game if the count reaches the count of 17 or higher but lesser than 21 ). Once the machine stops hitting, evaluation will be done to announce the winner.  If the Player won the game, the balance will be summed up with the bet amount or else the bet amount will be handed over to the machine if the Machine won the game :)